test/web-platform-tests/to-upstream/XMLHttpRequest/send-authentication-cors-post.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: send() - CORS POST basic authenticated request</title>
    <link rel="author" title="Maxim Vorobjov" href="mailto:maxim.vorobjov@gmail.com">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-setrequestheader()-method" data-tested-assertations="following::ol[1]/li[6]" />
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-send()-method" data-tested-assertations="following::code[contains(@title,'http-authorization')]/.." />
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var client = new XMLHttpRequest()

        client.open("POST", "http://www1.w3c-test.org/XMLHttpRequest/resources/auth7/corsenabled.py", true)
        client.setRequestHeader('Authorization', 'Basic MTMwZGY5NzEtMDE1NS00NTE4LWFjMjctOTU0ZmFhMzUwNWEzOnBhc3M=')
        client.setRequestHeader("x-user", "130df971-0155-4518-ac27-954faa3505a3")
        client.setRequestHeader("x-pass", "pass")
        client.setRequestHeader('Accept', 'text/plain');
        client.setRequestHeader('Content-type', 'application/json; charset=utf-8');
        client.onreadystatechange = function () {
          if (client.readyState < 4) {return}
          test.step( function () {
            assert_equals(client.status, 200)
            test.done()
          })
        }
        client.send("{'data':'test'}")
      })
    </script>
  </body>
</html>
